Moving fnc to look for segmentation cell to DetUtils
