+++ Detector/DetCommon/CMakeLists.txt
@@ -5,7 +5,7 @@ gaudi_subdir(DetCommon v1r0)
 
 gaudi_depends_on_subdirs(GaudiKernel)
 
-find_package(DD4hep)
+find_package(DD4hep COMPONENTS DDG4 REQUIRED)
 find_package(Geant4)
 include(${Geant4_USE_FILE})
 
@@ -18,7 +18,8 @@ gaudi_install_headers(DetCommon)
 gaudi_add_library(DetCommon
                  src/*.cpp
                  INCLUDE_DIRS DD4hep ROOT Geant4
-                 LINK_LIBRARIES GaudiKernel DD4hep ROOT Geant4)
+                 LINK_LIBRARIES GaudiKernel DD4hep ROOT Geant4 ${DD4hep_COMPONENT_LIBRARIES}
+                 PUBLIC_HEADERS DetCommon)
 
 set(LIBRARY_OUTPUT_PATH ${CMAKE_LIBRARY_OUTPUT_DIRECTORY})
 message(STATUS "LIBRARY_OUTPUT_PATH -> ${LIBRARY_OUTPUT_PATH}")
