+++ Detector/DetSensitive/CMakeLists.txt
@@ -3,7 +3,7 @@
 ################################################################################
 gaudi_subdir(DetSensitive v1r0)
 
-gaudi_depends_on_subdirs(GaudiKernel)
+gaudi_depends_on_subdirs(GaudiKernel Detector/DetCommon)
 
 find_package(DD4hep COMPONENTS DDG4 REQUIRED)
 find_package(Geant4)
@@ -16,8 +16,8 @@ gaudi_install_headers(DetSensitive)
 
 gaudi_add_library(DetSensitive
                   src/*.cpp
-                  INCLUDE_DIRS Geant4 DD4hep ROOT
-                  LINK_LIBRARIES GaudiKernel Geant4 DD4hep ROOT ${DD4hep_COMPONENT_LIBRARIES}
+                  INCLUDE_DIRS Geant4 DD4hep ROOT DetCommon
+                  LINK_LIBRARIES GaudiKernel DetCommon Geant4 DD4hep ROOT ${DD4hep_COMPONENT_LIBRARIES}
                   PUBLIC_HEADERS DetSensitive)
 
 set(CMAKE_MODULE_PATH  ${CMAKE_MODULE_PATH}  ${DD4hep_ROOT}/cmake )
