+++ Detector/DetSensitive/src/SimpleTrackerSD.cpp
@@ -1,7 +1,7 @@
 #include "DetSensitive/SimpleTrackerSD.h"
 
 // FCCSW
-#include "DetSensitive/SegmentationHelper.h"
+#include "DetCommon/DetUtils.h"
 
 // DD4hep
 #include "DDG4/Geant4Mapping.h"
@@ -49,7 +49,7 @@ bool SimpleTrackerSD::ProcessHits(G4Step* aStep, G4TouchableHistory*)
   // deleted in ~G4Event
   auto hit = new DD4hep::Simulation::Geant4TrackerHit(
     track->GetTrackID(), track->GetDefinition()->GetPDGEncoding(),edep, track->GetGlobalTime());
-  hit->cellID  = segmentation::cellID(m_seg, *aStep);
+  hit->cellID  = utils::cellID(m_seg, *aStep);
   hit->energyDeposit = edep;
   hit->position = position;
   hit->momentum = direction;
